Special SourceLinkContainer
The LinkContainer for Sources requires a slightly different append()
function, so we override the LinkContainer base class.
