+++ nixio/entity.py
@@ -8,8 +8,6 @@
 # LICENSE file in the root of the Project.
 
 from . import util
-from .source import Source
-from .container import LinkContainer
 
 
 class Entity(object):
@@ -140,20 +138,6 @@ class Entity(object):
         util.check_attr_type(t, str)
         self._h5group.set_attr("type", t)
 
-    @property
-    def sources(self):
-        """
-        A property containing all Sources referenced by the group. Sources
-        can be obtained by index or their id. Sources can be removed from the
-        list, but removing a referenced Source will not remove it from the
-        file. New Sources can be added using the append method of the list.
-        This is a read only attribute.
-        """
-        if self._sources is None:
-            self._sources = SourceLinkContainer("sources", self, Source,
-                                                self._parent.sources)
-        return self._sources
-
     def __eq__(self, other):
         if hasattr(other, "id"):
             return self.id == other.id
@@ -170,23 +154,3 @@ class Entity(object):
 
     def __repr__(self):
         return self.__str__()
-
-
-class SourceLinkContainer(LinkContainer):
-
-    # TODO: Sources returned from this container have an incorrect _parent
-    # This is the same issue that we have with Sections. It should probably be
-    # solved the same way
-
-    def append(self, item):
-        if util.is_uuid(item):
-            item = self._inst_item(self._backend.get_by_id(item))
-
-        if not hasattr(item, "id"):
-            raise TypeError("NIX entity or id string required for append")
-
-        if not self._itemstore._parent.find_sources(filtr=lambda x:
-                                                    x.id == item.id):
-            raise RuntimeError("This item cannot be appended here.")
-
-        self._backend.create_link(item, item.id)
