+++ nixio/group.py
@@ -73,3 +73,16 @@ class Group(Entity):
             self._multi_tags = LinkContainer("multi_tags", self, MultiTag,
                                              self._parent.multi_tags)
         return self._multi_tags
+
+    @property
+    def sources(self):
+        """
+        A property containing all Sources referenced by the Group. Sources
+        can be obtained by index or their id. Sources can be removed from the
+        list, but removing a referenced Source will not remove it from the
+        file. New Sources can be added using the append method of the list.
+        This is a read only attribute.
+        """
+        if self._sources is None:
+            self._sources = SourceLinkContainer(self)
+        return self._sources
